Add pinmux helper for A15 pin
By default, this provides a clock used for CEC by the HDMI framer,
but there is an internal clock and almost no one uses CEC. A lot of
EMC noise is made if this signal is left on, so it should be turned
off. This provides the ability to turn it off.